requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-grip==3.5.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jwcrypto==1.5.6
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.4.0
pubcontrol==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zope.interface==7.0.1
